Share styles between speakers and teams
