+++ src/components/Presenters/Row.module.scss
@@ -1,7 +1,7 @@
 @import '../../utils/breakpoints';
 @import '../../utils/variables';
 
-.speakerContainer {
+.presenterContainer {
   position: relative;
   display: flex;
   flex-wrap: nowrap;
@@ -39,7 +39,7 @@
   }
 }
 
-.speakerPicture {
+.presenterPicture {
   text-align: center;
   flex-grow: 0;
   flex-shrink: 0;
@@ -58,7 +58,7 @@
   }
 }
 
-.speakerInfo {
+.presenterInfo {
   flex-grow: 1;
   flex-shrink: 1;
   padding: 0 0.5em;
