+++ src/components/Presenters/Speaker/Page.jsx
@@ -10,15 +10,15 @@ import Img from 'gatsby-image';
 import Overdrive from 'react-overdrive';
 
 // Styles
-import styles from './Page.module.scss';
+import styles from '../Page.module.scss';
 
 // Components
-import EventRow from '../Event/Row';
+import EventRow from '../../Event/Row';
 
 // Images
-import twitterLogo from './twitter.png';
+import twitterLogo from '../twitter.png';
 
-const Page = ({ team, events, children }) => {
+const Page = ({ speaker, events, children }) => {
   const sessions =
     events && events.length && events.filter(e => e).length ? (
       <Fragment>
@@ -48,26 +48,32 @@ const Page = ({ team, events, children }) => {
   return (
     <section className={styles.container}>
       <div className={styles.header}>
-        <Overdrive id={`${team.teamName}-pic`} className={styles.headerPicture}>
-          {team.picture.type === Img
-            ? React.cloneElement(team.picture, {
-                alt: `${team.teamName}`,
+        <Overdrive
+          id={`${speaker.firstname}-${speaker.lastname}-pic`}
+          className={styles.headerPicture}
+        >
+          {speaker.picture.type === Img
+            ? React.cloneElement(speaker.picture, {
+                alt: `${speaker.firstname} ${speaker.lastname}`,
               })
-            : React.cloneElement(team.picture, {
-                alt: `${team.teamName}`,
+            : React.cloneElement(speaker.picture, {
+                alt: `${speaker.firstname} ${speaker.lastname}`,
               })}
         </Overdrive>
         <div className={styles.headerTitles}>
-          <h3>{team.teamName}</h3>
-          <h4 className={styles.title}>{team.intro}</h4>
-          {team.twitter ? (
+          <h3>
+            {speaker.firstname}{' '}
+            <span className={styles.lastname}>{speaker.lastname}</span>
+          </h3>
+          <h4 className={styles.title}>{speaker.title}</h4>
+          {speaker.twitter ? (
             <a
               className={styles.twitter}
-              href={`https://twitter.com/${team.twitter.substr(1)}`}
+              href={`https://twitter.com/${speaker.twitter.substr(1)}`}
               target="_blank"
               rel="noopener noreferrer"
             >
-              <img src={twitterLogo} alt="Twitter Feed" /> {team.twitter}
+              <img src={twitterLogo} alt="Twitter Feed" /> {speaker.twitter}
             </a>
           ) : null}
         </div>
@@ -79,13 +85,13 @@ const Page = ({ team, events, children }) => {
 };
 
 Page.propTypes = {
-  team: PropTypes.object,
+  speaker: PropTypes.object,
   events: PropTypes.array,
   children: PropTypes.node,
 };
 
 Page.defaultProps = {
-  team: {},
+  speaker: {},
   events: [],
 };
