+++ src/components/Presenters/Speaker/Row.jsx
@@ -10,11 +10,11 @@ import Link from 'gatsby-link';
 import Img from 'gatsby-image';
 import Overdrive from 'react-overdrive';
 
-import styles from './Row.module.scss';
+import styles from '../Row.module.scss';
 
 const Row = ({ speaker }) => (
   <Link
-    className={styles.speakerContainer}
+    className={styles.presenterContainer}
     to={{
       pathname: speaker.slug,
       state: { modal: true },
@@ -23,7 +23,7 @@ const Row = ({ speaker }) => (
     {speaker.picture ? (
       <Overdrive
         id={`${speaker.firstname}-${speaker.lastname}-pic`}
-        className={styles.speakerPicture}
+        className={styles.presenterPicture}
       >
         <Img
           sizes={speaker.picture.childImageSharp.sizes}
@@ -33,7 +33,7 @@ const Row = ({ speaker }) => (
     ) : (
       ''
     )}
-    <div className={styles.speakerInfo}>
+    <div className={styles.presenterInfo}>
       <h3>
         {speaker.firstname}{' '}
         <span className={styles.lastname}>{speaker.lastname}</span>
