--- src/components/Team/Card.module.scss
@@ -1,100 +1,0 @@
-@import '../../utils/breakpoints';
-@import '../../utils/variables';
-
-.item {
-  width: 12em;
-  max-width: 14em;
-  text-align: center;
-  margin: 0.7em;
-  flex-grow: 1;
-  flex-shrink: 0;
-  flex-basis: 12em;
-  text-decoration: none;
-  border-radius: 2px;
-  background-color: lighten($background-grey, 2%);
-  box-shadow: 0 1px 2px 0 rgba(0, 0, 0, 0.1);
-  position: relative;
-
-  &::before {
-    content: '';
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: #fff;
-    box-shadow: 0 1px 2px 0 rgba(0, 0, 0, 0.2);
-    opacity: 0;
-    transition: opacity 0.3s ease-in-out;
-    will-change: opacity;
-  }
-
-  &:hover,
-  &:focus,
-  &:active {
-    text-decoration: none;
-
-    &::before {
-      opacity: 1;
-    }
-  }
-
-  @include mappy-bp(medium) {
-    flex-grow: 0;
-  }
-}
-
-.info {
-  padding: 1em;
-}
-
-.name {
-  position: relative;
-  display: block;
-  color: $ict-grey-text;
-  text-decoration: none;
-  font-size: 1.1em;
-  line-height: 1.4;
-
-  .item:hover &,
-  .item:focus &,
-  .item:active & {
-    color: darken($ict-grey-text, 15%);
-  }
-}
-
-.title {
-  position: relative;
-  margin-top: 0.8em;
-  color: $ict-blue-text;
-  font-size: 0.85em;
-  line-height: 1.4;
-}
-
-.pictureFrame {
-  position: relative;
-  line-height: 0;
-  width: 100%;
-  display: inline-block;
-  overflow: hidden;
-}
-
-.picture {
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  width: 100%;
-  height: 100%;
-  transform-origin: center;
-  transition: all 250ms ease-in-out;
-  will-change: opacity, transform;
-
-  .item:hover &,
-  .item:focus &,
-  .item:active & {
-    opacity: 0.9;
-    transform: scale(1.1);
-  }
-}
