--- src/components/Team/Row.module.scss
@@ -1,90 +1,0 @@
-@import '../../utils/breakpoints';
-@import '../../utils/variables';
-
-.speakerContainer {
-  position: relative;
-  display: flex;
-  flex-wrap: nowrap;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  border: 1px solid #e7e7e7;
-  border-radius: 2px;
-  background-color: #fff;
-  margin-bottom: 1em;
-  text-decoration: none;
-
-  // Hardware accelerated animation of box-shadow (see http://tobiasahlin.com/blog/how-to-animate-box-shadow/)
-  &::before {
-    content: '';
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    box-shadow: 0 2px 2px 0 rgba(0, 0, 0, 0.2), 0 3px 3px 0 rgba(0, 0, 0, 0.22);
-    opacity: 0;
-    transition: opacity 0.3s ease-in-out;
-    will-change: opacity;
-  }
-
-  &:hover,
-  &:focus,
-  &:active {
-    text-decoration: none;
-
-    &::before {
-      opacity: 1;
-    }
-  }
-}
-
-.speakerPicture {
-  text-align: center;
-  flex-grow: 0;
-  flex-shrink: 0;
-  flex-basis: 7em;
-  width: 7em;
-  background-color: #000;
-
-  img {
-    display: block;
-    width: 100%;
-  }
-
-  @include mappy-bp(medium) {
-    flex-basis: 10em;
-    width: 10em;
-  }
-}
-
-.speakerInfo {
-  flex-grow: 1;
-  flex-shrink: 1;
-  padding: 0 0.5em;
-
-  h3 {
-    font-size: 1.2em;
-    margin: 0;
-    color: $ict-grey-text;
-  }
-
-  @include mappy-bp(medium) {
-    padding: 0 1em;
-
-    h3 {
-      font-size: 1.3em;
-    }
-  }
-}
-
-.title {
-  font-size: 0.9em;
-  font-weight: 400;
-  margin: 0;
-  color: $ict-blue-text;
-
-  @include mappy-bp(medium) {
-    font-size: 1em;
-  }
-}
